temp

+++ Handbook/Models/Law.swift
@@ -24,10 +24,10 @@ struct TextContent : Identifiable {
     
     class Content: Identifiable{
         var id: UUID = UUID()
-        var line: Int
+        var line: Int64
         var text: String
         
-        init(_ line: Int, _ text: String) {
+        init(_ line: Int64, _ text: String) {
             self.line = line
             self.text = text
         }
@@ -36,7 +36,7 @@ struct TextContent : Identifiable {
     var id: UUID = UUID()
     var text: String
     var children: [Content] = []
-    var line: Int
+    var line: Int64
     var indent: Int
 }
 
@@ -45,9 +45,9 @@ class TocListData: Identifiable {
     var children: [TocListData] = []
     var title: String
     var indent: Int
-    var line: Int
+    var line: Int64
 
-    init(title: String, indent: Int, line: Int){
+    init(title: String, indent: Int, line: Int64){
         self.title = title
         self.indent = indent
         self.line = line
